+++ src/file.h
@@ -21,8 +21,8 @@ class File : public FileInterface {
   virtual Status Open(const Mode mode);
   virtual Status Close();
   virtual Status Unlink();
-  virtual int32_t Tell();
-  virtual Status Seek(int32_t offset);
+  virtual int64_t Tell();
+  virtual Status Seek(int64_t offset);
   virtual Status SeekEof();
   virtual Status Read(void* buffer, size_t length, size_t* read_bytes);
   virtual Status ReadLines(std::vector<std::string>* strings);
